+++ backend/chat/src/main/java/com/example/ChatApplication.java
@@ -1,4 +1,4 @@
-package com.example.chat;
+package com.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
