+++ backend/chat/src/main/java/com/example/config/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package com.example.chat.config;
+package com.example.config;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
