+++ backend/chat/src/main/java/com/example/repository/TopicRepository.java
@@ -1,6 +1,6 @@
-package com.example.chat.repository;
+package com.example.repository;
 
-import com.example.chat.domain.Topic;
+import com.example.domain.Topic;
 import org.springframework.data.repository.CrudRepository;
 
 public interface TopicRepository extends CrudRepository<Topic, String> {
